# Implement `waiting_trials_cache` in study