--- a/optuna/study/study.py
+++ b/optuna/study/study.py
@@ -94,7 +94,7 @@ class Study:
 
         self._thread_local = _ThreadLocalStudyAttribute()
         self._stop_flag = False
-        self._waiting_trials_cache: set[FrozenTrial] | None = None
+        self._waiting_trials_cache: dict = {}
 
     def __getstate__(self) -> dict[Any, Any]:
         state = self.__dict__.copy()
@@ -1057,37 +1057,32 @@ class Study:
         return len(self.directions) > 1
 
     def _pop_waiting_trial_id(self) -> int | None:
-        if self._waiting_trials_cache is not None:
-            # trials_to_be_removed = set()
-            for trial in self._waiting_trials_cache:
-                # trials_to_be_removed.add(trial)
-                self._waiting_trials_cache -= {trial}
-                if not self._storage.set_trial_state_values(
-                    trial._trial_id, state=TrialState.RUNNING
-                ):
+        if self._waiting_trials_cache:
+            for trial_id, trial_number in self._waiting_trials_cache.items():
+                del self._waiting_trials_cache[trial_id]
+                if not self._storage.set_trial_state_values(trial_id, state=TrialState.RUNNING):
                     continue
 
-                _logger.debug("Trial {} popped from the trial queue.".format(trial.number))
-                # self._waiting_trials_cache -= trials_to_be_removed
-                return trial._trial_id
+                _logger.debug("Trial {} popped from the trial queue.", trial_number)
+                return trial_id
 
         waiting_trials = self._storage.get_all_trials(
             self._study_id, deepcopy=False, states=(TrialState.WAITING,)
         )
-        trials_to_be_removed = set()
+        self._waiting_trials_cache = {}
+        poped_trial_id = None
         for trial in waiting_trials:
-            trials_to_be_removed.add(trial)
-            if not self._storage.set_trial_state_values(trial._trial_id, state=TrialState.RUNNING):
-                continue
+            if poped_trial_id is None:
+                if not self._storage.set_trial_state_values(
+                    trial._trial_id, state=TrialState.RUNNING
+                ):
+                    continue
+                poped_trial_id = trial._trial_id
+                _logger.debug("Trial {} popped from the trial queue.", trial.number)
+            else:
+                self._waiting_trials_cache[trial._trial_id] = trial.number
 
-            _logger.debug("Trial {} popped from the trial queue.".format(trial.number))
-            self._waiting_trials_cache = set(
-                trial for trial in waiting_trials if trial not in trials_to_be_removed
-            )
-            return trial._trial_id
-
-        self._waiting_trials_cache = None
-        return None
+        return poped_trial_id
 
     def _should_skip_enqueue(self, params: Mapping[str, JSONSerializable]) -> bool:
         for trial in self.get_trials(deepcopy=False):